requirement.txt
torch==2.1.2
numpy==1.26.3
pyarrow==14.0.2
torchvision==0.16.2
opencv-python==4.9.0.80
pyyaml==6.0.1
submitit==1.5.1
tqdm==4.66.1
